requirements.txt
langchain
langchain-community
python-dotenv
streamlit
openai
faiss-cpu
tiktoken
unstructured
